refactor: raffle list
