+++ src/pages/LiquidationsPage/components/GraceCard/index.ts
@@ -1,1 +1,1 @@
-export * from './GraceCard';
+export { default } from './GraceCard';
