+++ src/pages/LiquidationsPage/components/Liquidations/Liquidations.module.scss
@@ -9,7 +9,3 @@
   padding: 24px;
   min-height: 600px;
 }
-
-.container {
-  position: relative;
-}
