+++ src/pages/LiquidationsPage/components/OngoingRaffleTab/index.ts
@@ -1,0 +1,1 @@
+export { default } from './OngoingRaffleTab';
