+++ src/pages/LiquidationsPage/components/UpcomingRaffleTab/index.ts
@@ -1,0 +1,1 @@
+export { default } from './UpcomingRaffleTab';
