+++ src/pages/LiquidationsPage/components/WonRaffleTab/index.ts
@@ -1,0 +1,1 @@
+export { default } from './WonRaffleTab';
